requirements.txt
langchain==0.2.10
langchain-community==0.2.10
langchain-openai==0.1.7
langchain-chroma==0.1.2
faiss-cpu==1.9.0.post1
pypdf==4.2.0
python-dotenv==1.0.1
streamlit==1.36.0